% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raging_bool.R
\name{raging_bool}
\alias{raging_bool}
\alias{print.raging_bool}
\alias{!.raging_bool}
\alias{&.raging_bool}
\alias{|.raging_bool}
\alias{as.logical.raging_bool}
\alias{is.raging_bool}
\alias{[.raging_bool}
\alias{c.raging_bool}
\title{Raging boolean class}
\usage{
raging_bool(x = logical())

\method{print}{raging_bool}(x, ...)

\method{!}{raging_bool}(x)

\method{&}{raging_bool}(e1, e2)

\method{|}{raging_bool}(e1, e2)

\method{as.logical}{raging_bool}(x, ...)

is.raging_bool(x)

\method{[}{raging_bool}(x, i)

\method{c}{raging_bool}(..., recursive = FALSE, use.names = TRUE)
}
\arguments{
\item{x}{logical vector}
}
\value{
object of class 'raging_bool'
}
\description{
An object of class 'raging_bool' is logical vector with a 'rage' attribute.
Based on this attribute, values can randomly, silently (and mercilessly!) be
inverted when working with or printing them.
}
\examples{
a <- raging_bool(c(TRUE, FALSE))
a
}
